ai_service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.27.0
PyPDF2>=3.0.0
